+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.0] - 2022-02-06
+
+### Added
+
+* A dedicated `migrate` command, distinct from `restore`.
+
 ## [4.0.8] - 2021-12-23
 
 ### Added